add model handler
When a manifest is received from Golioth that contains a package named
`model` it will be used as the selected TensorFlow model. The model will be
downloaded and saved to the SD card if not already present. The model will
then be loaded into memory from the SD card, parsing the custom header to
collect the number of categories in the training set and a text label for
each of those categories.

Signed-off-by: Mike Szczys <mike@golioth.io>