store the choice of model on the SD card
Several models may be downloaded to the SD card. By storing the path to the
most recently chosen model, the device may start recognition before
receiving the manifest from the server.

A reboot has been added when a new model choice is received. This is the
easiest way to free an already loaded model and load the new one. Because
the choice is now stored locally, the reboot doesn't need to wait for a
manifest before loading.

Signed-off-by: Mike Szczys <mike@golioth.io>